+++ x/global/keeper/msg_server_update_params.go
@@ -13,17 +13,23 @@ import (
 	govTypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 )
 
-func (k msgServer) UpdateParams(goCtx context.Context, req *types.MsgUpdateParams) (*types.MsgUpdateParamsResponse, error) {
-	if k.authority != req.Authority {
-		return nil, errors.Wrapf(govTypes.ErrInvalidSigner, "invalid authority; expected %s, got %s", k.authority, req.Authority)
+func (k msgServer) UpdateParams(goCtx context.Context, msg *types.MsgUpdateParams) (*types.MsgUpdateParamsResponse, error) {
+	if k.authority != msg.Authority {
+		return nil, errors.Wrapf(govTypes.ErrInvalidSigner, "invalid authority; expected %s, got %s", k.authority, msg.Authority)
 	}
 
 	ctx := sdk.UnwrapSDKContext(goCtx)
-	params := k.GetParams(ctx)
+	oldParams := k.GetParams(ctx)
 
-	payload := params
-	_ = json.Unmarshal([]byte(req.Payload), &payload)
-	k.SetParams(ctx, payload)
+	newParams := oldParams
+	_ = json.Unmarshal([]byte(msg.Payload), &newParams)
+	k.SetParams(ctx, newParams)
+
+	_ = ctx.EventManager().EmitTypedEvent(&types.EventUpdateParams{
+		OldParams: oldParams,
+		NewParams: newParams,
+		Payload:   msg.Payload,
+	})
 
 	return &types.MsgUpdateParamsResponse{}, nil
 }