+++ x/query/types/keys.go
@@ -7,13 +7,9 @@ const (
 	// StoreKey defines the primary module store key
 	StoreKey = ModuleName
 
-	// RouterKey is the message route for slashing
+	// RouterKey is the message route for query
 	RouterKey = ModuleName
 
 	// MemStoreKey defines the in-memory store key
 	MemStoreKey = "mem_query"
 )
-
-func KeyPrefix(p string) []byte {
-	return []byte(p)
-}