+++ x/stakers/types/events.pb.go
@@ -5,6 +5,8 @@ package types
 
 import (
 	fmt "fmt"
+	github_com_cosmos_cosmos_sdk_types "github.com/cosmos/cosmos-sdk/types"
+	_ "github.com/cosmos/gogoproto/gogoproto"
 	proto "github.com/gogo/protobuf/proto"
 	io "io"
 	math "math"
@@ -22,6 +24,71 @@ var _ = math.Inf
 // proto package needs to be updated.
 const _ = proto.GoGoProtoPackageIsVersion3 // please upgrade the proto package
 
+// EventUpdateParams is an event emitted when the module parameters are updated.
+// emitted_by: MsgUpdateParams
+type EventUpdateParams struct {
+	// old_params is the module's old parameters.
+	OldParams Params `protobuf:"bytes,1,opt,name=old_params,json=oldParams,proto3" json:"old_params"`
+	// new_params is the module's new parameters.
+	NewParams Params `protobuf:"bytes,2,opt,name=new_params,json=newParams,proto3" json:"new_params"`
+	// payload is the parameter updates that were performed.
+	Payload string `protobuf:"bytes,3,opt,name=payload,proto3" json:"payload,omitempty"`
+}
+
+func (m *EventUpdateParams) Reset()         { *m = EventUpdateParams{} }
+func (m *EventUpdateParams) String() string { return proto.CompactTextString(m) }
+func (*EventUpdateParams) ProtoMessage()    {}
+func (*EventUpdateParams) Descriptor() ([]byte, []int) {
+	return fileDescriptor_7a1b3dc9634155a0, []int{0}
+}
+func (m *EventUpdateParams) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *EventUpdateParams) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_EventUpdateParams.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *EventUpdateParams) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_EventUpdateParams.Merge(m, src)
+}
+func (m *EventUpdateParams) XXX_Size() int {
+	return m.Size()
+}
+func (m *EventUpdateParams) XXX_DiscardUnknown() {
+	xxx_messageInfo_EventUpdateParams.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_EventUpdateParams proto.InternalMessageInfo
+
+func (m *EventUpdateParams) GetOldParams() Params {
+	if m != nil {
+		return m.OldParams
+	}
+	return Params{}
+}
+
+func (m *EventUpdateParams) GetNewParams() Params {
+	if m != nil {
+		return m.NewParams
+	}
+	return Params{}
+}
+
+func (m *EventUpdateParams) GetPayload() string {
+	if m != nil {
+		return m.Payload
+	}
+	return ""
+}
+
 // EventCreateStaker is an event emitted when a protocol node stakes in a pool.
 // emitted_by: MsgCreateStaker
 type EventCreateStaker struct {
@@ -35,7 +102,7 @@ func (m *EventCreateStaker) Reset()         { *m = EventCreateStaker{} }
 func (m *EventCreateStaker) String() string { return proto.CompactTextString(m) }
 func (*EventCreateStaker) ProtoMessage()    {}
 func (*EventCreateStaker) Descriptor() ([]byte, []int) {
-	return fileDescriptor_7a1b3dc9634155a0, []int{0}
+	return fileDescriptor_7a1b3dc9634155a0, []int{1}
 }
 func (m *EventCreateStaker) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
@@ -95,7 +162,7 @@ func (m *EventUpdateMetadata) Reset()         { *m = EventUpdateMetadata{} }
 func (m *EventUpdateMetadata) String() string { return proto.CompactTextString(m) }
 func (*EventUpdateMetadata) ProtoMessage()    {}
 func (*EventUpdateMetadata) Descriptor() ([]byte, []int) {
-	return fileDescriptor_7a1b3dc9634155a0, []int{1}
+	return fileDescriptor_7a1b3dc9634155a0, []int{2}
 }
 func (m *EventUpdateMetadata) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
@@ -158,14 +225,14 @@ type EventUpdateCommission struct {
 	// staker is the account address of the protocol node.
 	Staker string `protobuf:"bytes,1,opt,name=staker,proto3" json:"staker,omitempty"`
 	// commission ...
-	Commission string `protobuf:"bytes,2,opt,name=commission,proto3" json:"commission,omitempty"`
+	Commission github_com_cosmos_cosmos_sdk_types.Dec `protobuf:"bytes,2,opt,name=commission,proto3,customtype=github.com/cosmos/cosmos-sdk/types.Dec" json:"commission"`
 }
 
 func (m *EventUpdateCommission) Reset()         { *m = EventUpdateCommission{} }
 func (m *EventUpdateCommission) String() string { return proto.CompactTextString(m) }
 func (*EventUpdateCommission) ProtoMessage()    {}
 func (*EventUpdateCommission) Descriptor() ([]byte, []int) {
-	return fileDescriptor_7a1b3dc9634155a0, []int{2}
+	return fileDescriptor_7a1b3dc9634155a0, []int{3}
 }
 func (m *EventUpdateCommission) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
@@ -201,13 +268,6 @@ func (m *EventUpdateCommission) GetStaker() string {
 	return ""
 }
 
-func (m *EventUpdateCommission) GetCommission() string {
-	if m != nil {
-		return m.Commission
-	}
-	return ""
-}
-
 // EventJoinPool ...
 // emitted_by: MsgJoinPool
 type EventJoinPool struct {
@@ -226,7 +286,7 @@ func (m *EventJoinPool) Reset()         { *m = EventJoinPool{} }
 func (m *EventJoinPool) String() string { return proto.CompactTextString(m) }
 func (*EventJoinPool) ProtoMessage()    {}
 func (*EventJoinPool) Descriptor() ([]byte, []int) {
-	return fileDescriptor_7a1b3dc9634155a0, []int{3}
+	return fileDescriptor_7a1b3dc9634155a0, []int{4}
 }
 func (m *EventJoinPool) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
@@ -296,7 +356,7 @@ func (m *EventLeavePool) Reset()         { *m = EventLeavePool{} }
 func (m *EventLeavePool) String() string { return proto.CompactTextString(m) }
 func (*EventLeavePool) ProtoMessage()    {}
 func (*EventLeavePool) Descriptor() ([]byte, []int) {
-	return fileDescriptor_7a1b3dc9634155a0, []int{4}
+	return fileDescriptor_7a1b3dc9634155a0, []int{5}
 }
 func (m *EventLeavePool) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
@@ -340,6 +400,7 @@ func (m *EventLeavePool) GetStaker() string {
 }
 
 func init() {
+	proto.RegisterType((*EventUpdateParams)(nil), "kyve.stakers.v1beta1.EventUpdateParams")
 	proto.RegisterType((*EventCreateStaker)(nil), "kyve.stakers.v1beta1.EventCreateStaker")
 	proto.RegisterType((*EventUpdateMetadata)(nil), "kyve.stakers.v1beta1.EventUpdateMetadata")
 	proto.RegisterType((*EventUpdateCommission)(nil), "kyve.stakers.v1beta1.EventUpdateCommission")
@@ -350,29 +411,86 @@ func init() {
 func init() { proto.RegisterFile("kyve/stakers/v1beta1/events.proto", fileDescriptor_7a1b3dc9634155a0) }
 
 var fileDescriptor_7a1b3dc9634155a0 = []byte{
-	// 342 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x9c, 0x92, 0xcd, 0x4a, 0xeb, 0x40,
-	0x14, 0xc7, 0x9b, 0xde, 0xd0, 0x72, 0x07, 0xee, 0x85, 0x9b, 0xeb, 0x47, 0x56, 0x43, 0xcd, 0xaa,
-	0x0b, 0x49, 0x28, 0x3e, 0x81, 0x96, 0x0a, 0x7e, 0x4b, 0x45, 0x41, 0x37, 0x32, 0x69, 0x0e, 0xed,
-	0xd0, 0x24, 0x27, 0x64, 0x4e, 0xd3, 0xf6, 0x2d, 0x7c, 0x2c, 0x97, 0x5d, 0xba, 0x94, 0xf6, 0x45,
-	0x24, 0xd3, 0xb1, 0x44, 0xa1, 0x1b, 0x77, 0xf9, 0x7f, 0xf0, 0xff, 0x91, 0xe1, 0xb0, 0x83, 0xf1,
-	0xbc, 0x80, 0x40, 0x91, 0x18, 0x43, 0xae, 0x82, 0xa2, 0x13, 0x02, 0x89, 0x4e, 0x00, 0x05, 0xa4,
-	0xa4, 0xfc, 0x2c, 0x47, 0x42, 0x67, 0xa7, 0xac, 0xf8, 0xa6, 0xe2, 0x9b, 0x8a, 0xd7, 0x65, 0xff,
-	0x7a, 0x65, 0xab, 0x9b, 0x83, 0x20, 0xb8, 0xd3, 0xa9, 0xb3, 0xc7, 0x1a, 0xeb, 0x9e, 0x6b, 0xb5,
-	0xac, 0xf6, 0xef, 0xbe, 0x51, 0xa5, 0x2f, 0x12, 0x9c, 0xa4, 0xe4, 0xd6, 0x5b, 0x56, 0xdb, 0xee,
-	0x1b, 0xe5, 0x4d, 0xd8, 0x7f, 0x3d, 0x72, 0x9f, 0x45, 0x82, 0xe0, 0x0a, 0x48, 0x44, 0x82, 0xc4,
-	0xd6, 0x19, 0x97, 0x35, 0x13, 0x4c, 0x65, 0x19, 0xd4, 0x75, 0xf0, 0x29, 0xcb, 0x64, 0x0a, 0xa1,
-	0x92, 0x04, 0xee, 0xaf, 0x75, 0x62, 0xa4, 0xe3, 0x30, 0x3b, 0xc6, 0x21, 0xba, 0xb6, 0xb6, 0xf5,
-	0xb7, 0x77, 0xc3, 0x76, 0x2b, 0xd8, 0x2e, 0x26, 0x89, 0x54, 0x4a, 0x62, 0xba, 0x15, 0xcc, 0x19,
-	0x1b, 0x6c, 0x5a, 0x86, 0x5d, 0x71, 0xbc, 0x19, 0xfb, 0xa3, 0x07, 0xcf, 0x51, 0xa6, 0xb7, 0x88,
-	0xb1, 0xb3, 0xcf, 0x9a, 0x19, 0x62, 0xfc, 0x2c, 0x23, 0xbd, 0x64, 0xf7, 0x1b, 0xa5, 0x3c, 0x8b,
-	0x2a, 0x84, 0xfa, 0x77, 0x42, 0x21, 0x62, 0x11, 0x45, 0x39, 0x28, 0x65, 0xfe, 0xa1, 0xe2, 0x54,
-	0x5e, 0xd0, 0xfe, 0xf2, 0x82, 0xc7, 0xec, 0xaf, 0x26, 0x5f, 0x82, 0x28, 0xe0, 0x47, 0xe8, 0x93,
-	0xd3, 0xd7, 0x25, 0xb7, 0x16, 0x4b, 0x6e, 0xbd, 0x2f, 0xb9, 0xf5, 0xb2, 0xe2, 0xb5, 0xc5, 0x8a,
-	0xd7, 0xde, 0x56, 0xbc, 0xf6, 0x74, 0x38, 0x94, 0x34, 0x9a, 0x84, 0xfe, 0x00, 0x93, 0xe0, 0xe2,
-	0xf1, 0xa1, 0x77, 0x0d, 0x34, 0xc5, 0x7c, 0x1c, 0x0c, 0x46, 0x42, 0xa6, 0xc1, 0x6c, 0x73, 0x36,
-	0x34, 0xcf, 0x40, 0x85, 0x0d, 0x7d, 0x2e, 0x47, 0x1f, 0x01, 0x00, 0x00, 0xff, 0xff, 0xdf, 0x2b,
-	0x0f, 0x0f, 0x53, 0x02, 0x00, 0x00,
+	// 462 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x9c, 0x93, 0xdf, 0x6a, 0xd4, 0x40,
+	0x14, 0xc6, 0x37, 0x6b, 0xd8, 0x65, 0x8f, 0x28, 0x18, 0xab, 0x86, 0x22, 0x69, 0xcd, 0x85, 0xf4,
+	0x42, 0x13, 0xaa, 0x4f, 0xd0, 0x5d, 0x2b, 0xf8, 0xaf, 0x94, 0x88, 0x82, 0xde, 0x94, 0xb3, 0x99,
+	0xc3, 0x36, 0x6c, 0x92, 0x13, 0x32, 0xb3, 0xd9, 0xee, 0x95, 0xaf, 0xe0, 0xbb, 0xf8, 0x12, 0xbd,
+	0xec, 0xa5, 0x78, 0x51, 0x64, 0xf7, 0x45, 0x24, 0x93, 0xd9, 0x9a, 0x42, 0x17, 0xc4, 0xab, 0xcc,
+	0x37, 0xf3, 0xe5, 0xf7, 0x7d, 0x73, 0x48, 0xe0, 0xc9, 0x74, 0x51, 0x51, 0x28, 0x15, 0x4e, 0xa9,
+	0x94, 0x61, 0xb5, 0x3f, 0x26, 0x85, 0xfb, 0x21, 0x55, 0x94, 0x2b, 0x19, 0x14, 0x25, 0x2b, 0x76,
+	0xb6, 0x6a, 0x4b, 0x60, 0x2c, 0x81, 0xb1, 0x6c, 0x6f, 0x4d, 0x78, 0xc2, 0xda, 0x10, 0xd6, 0xab,
+	0xc6, 0xbb, 0x7d, 0x33, 0xae, 0xc0, 0x12, 0x33, 0x83, 0xf3, 0x7f, 0x58, 0x70, 0xef, 0xb0, 0xe6,
+	0x7f, 0x2a, 0x04, 0x2a, 0x3a, 0xd6, 0x67, 0xce, 0x01, 0x00, 0xa7, 0xe2, 0xa4, 0x71, 0xba, 0xd6,
+	0xae, 0xb5, 0x77, 0xfb, 0xc5, 0xe3, 0xe0, 0xa6, 0xe4, 0xa0, 0x79, 0x63, 0x68, 0x9f, 0x5f, 0xee,
+	0x74, 0xa2, 0x01, 0xa7, 0xe2, 0x2f, 0x22, 0xa7, 0xf9, 0x1a, 0xd1, 0xfd, 0x77, 0x44, 0x4e, 0x73,
+	0x83, 0x70, 0xa1, 0x5f, 0xe0, 0x22, 0x65, 0x14, 0xee, 0xad, 0x5d, 0x6b, 0x6f, 0x10, 0xad, 0xa5,
+	0x3f, 0x32, 0xa5, 0x47, 0x25, 0xa1, 0xa2, 0x8f, 0x9a, 0xe7, 0x3c, 0x84, 0x5e, 0x43, 0xd6, 0x85,
+	0x07, 0x91, 0x51, 0xf5, 0x3e, 0x66, 0x3c, 0xcb, 0x95, 0x6e, 0x61, 0x47, 0x46, 0xf9, 0x33, 0xb8,
+	0xdf, 0xba, 0xf9, 0x07, 0x52, 0x28, 0x50, 0xe1, 0x46, 0x8c, 0x0b, 0xfd, 0x8c, 0xf3, 0xa4, 0x3e,
+	0xe8, 0x36, 0x6d, 0x8c, 0xac, 0x4f, 0xe6, 0x34, 0x96, 0x89, 0xa2, 0x75, 0x4f, 0x23, 0x1d, 0x07,
+	0xec, 0x94, 0x27, 0xec, 0xda, 0x7a, 0x5b, 0xaf, 0xfd, 0x6f, 0xf0, 0xa0, 0x15, 0x3b, 0xe2, 0x2c,
+	0x4b, 0xa4, 0x4c, 0x38, 0xdf, 0x18, 0x7c, 0x04, 0x10, 0x5f, 0xb9, 0x9a, 0xec, 0x61, 0x50, 0xcf,
+	0xea, 0xd7, 0xe5, 0xce, 0xd3, 0x49, 0xa2, 0x4e, 0x67, 0xe3, 0x20, 0xe6, 0x2c, 0x8c, 0x59, 0x66,
+	0x2c, 0xcd, 0xe3, 0xb9, 0x14, 0xd3, 0x50, 0x2d, 0x0a, 0x92, 0xc1, 0x2b, 0x8a, 0xa3, 0x16, 0xc1,
+	0x3f, 0x83, 0x3b, 0xba, 0xc0, 0x5b, 0x4e, 0xf2, 0x63, 0xe6, 0xd4, 0x79, 0x04, 0xfd, 0x82, 0x39,
+	0x3d, 0x49, 0x84, 0x4e, 0xb6, 0xa3, 0x5e, 0x2d, 0xdf, 0x88, 0x56, 0xa3, 0xee, 0xb5, 0x46, 0x1e,
+	0x40, 0x85, 0x29, 0x0a, 0x51, 0x92, 0x94, 0xe6, 0xce, 0xad, 0x9d, 0xd6, 0xc4, 0xed, 0x6b, 0x13,
+	0x3f, 0x80, 0xbb, 0x3a, 0xf9, 0x3d, 0x61, 0x45, 0xff, 0x15, 0x3d, 0x7c, 0x7d, 0xbe, 0xf4, 0xac,
+	0x8b, 0xa5, 0x67, 0xfd, 0x5e, 0x7a, 0xd6, 0xf7, 0x95, 0xd7, 0xb9, 0x58, 0x79, 0x9d, 0x9f, 0x2b,
+	0xaf, 0xf3, 0xf5, 0x59, 0x6b, 0x14, 0xef, 0xbe, 0x7c, 0x3e, 0x3c, 0x22, 0x35, 0xe7, 0x72, 0x1a,
+	0xc6, 0xa7, 0x98, 0xe4, 0xe1, 0xd9, 0xd5, 0x6f, 0xa0, 0x87, 0x32, 0xee, 0xe9, 0xcf, 0xff, 0xe5,
+	0x9f, 0x00, 0x00, 0x00, 0xff, 0xff, 0x69, 0xd7, 0xac, 0xc1, 0x72, 0x03, 0x00, 0x00,
+}
+
+func (m *EventUpdateParams) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *EventUpdateParams) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *EventUpdateParams) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	if len(m.Payload) > 0 {
+		i -= len(m.Payload)
+		copy(dAtA[i:], m.Payload)
+		i = encodeVarintEvents(dAtA, i, uint64(len(m.Payload)))
+		i--
+		dAtA[i] = 0x1a
+	}
+	{
+		size, err := m.NewParams.MarshalToSizedBuffer(dAtA[:i])
+		if err != nil {
+			return 0, err
+		}
+		i -= size
+		i = encodeVarintEvents(dAtA, i, uint64(size))
+	}
+	i--
+	dAtA[i] = 0x12
+	{
+		size, err := m.OldParams.MarshalToSizedBuffer(dAtA[:i])
+		if err != nil {
+			return 0, err
+		}
+		i -= size
+		i = encodeVarintEvents(dAtA, i, uint64(size))
+	}
+	i--
+	dAtA[i] = 0xa
+	return len(dAtA) - i, nil
 }
 
 func (m *EventCreateStaker) Marshal() (dAtA []byte, err error) {
@@ -481,13 +599,16 @@ func (m *EventUpdateCommission) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 	_ = i
 	var l int
 	_ = l
-	if len(m.Commission) > 0 {
-		i -= len(m.Commission)
-		copy(dAtA[i:], m.Commission)
-		i = encodeVarintEvents(dAtA, i, uint64(len(m.Commission)))
-		i--
-		dAtA[i] = 0x12
+	{
+		size := m.Commission.Size()
+		i -= size
+		if _, err := m.Commission.MarshalTo(dAtA[i:]); err != nil {
+			return 0, err
+		}
+		i = encodeVarintEvents(dAtA, i, uint64(size))
 	}
+	i--
+	dAtA[i] = 0x12
 	if len(m.Staker) > 0 {
 		i -= len(m.Staker)
 		copy(dAtA[i:], m.Staker)
@@ -591,6 +712,23 @@ func encodeVarintEvents(dAtA []byte, offset int, v uint64) int {
 	dAtA[offset] = uint8(v)
 	return base
 }
+func (m *EventUpdateParams) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	l = m.OldParams.Size()
+	n += 1 + l + sovEvents(uint64(l))
+	l = m.NewParams.Size()
+	n += 1 + l + sovEvents(uint64(l))
+	l = len(m.Payload)
+	if l > 0 {
+		n += 1 + l + sovEvents(uint64(l))
+	}
+	return n
+}
+
 func (m *EventCreateStaker) Size() (n int) {
 	if m == nil {
 		return 0
@@ -642,10 +780,8 @@ func (m *EventUpdateCommission) Size() (n int) {
 	if l > 0 {
 		n += 1 + l + sovEvents(uint64(l))
 	}
-	l = len(m.Commission)
-	if l > 0 {
-		n += 1 + l + sovEvents(uint64(l))
-	}
+	l = m.Commission.Size()
+	n += 1 + l + sovEvents(uint64(l))
 	return n
 }
 
@@ -693,6 +829,154 @@ func sovEvents(x uint64) (n int) {
 }
 func sozEvents(x uint64) (n int) {
 	return sovEvents(uint64((x << 1) ^ uint64((int64(x) >> 63))))
+}
+func (m *EventUpdateParams) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowEvents
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: EventUpdateParams: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: EventUpdateParams: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		case 1:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field OldParams", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowEvents
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthEvents
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthEvents
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if err := m.OldParams.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field NewParams", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowEvents
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthEvents
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthEvents
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if err := m.NewParams.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		case 3:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Payload", wireType)
+			}
+			var stringLen uint64
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowEvents
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				stringLen |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			intStringLen := int(stringLen)
+			if intStringLen < 0 {
+				return ErrInvalidLengthEvents
+			}
+			postIndex := iNdEx + intStringLen
+			if postIndex < 0 {
+				return ErrInvalidLengthEvents
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.Payload = string(dAtA[iNdEx:postIndex])
+			iNdEx = postIndex
+		default:
+			iNdEx = preIndex
+			skippy, err := skipEvents(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
+				return ErrInvalidLengthEvents
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
 }
 func (m *EventCreateStaker) Unmarshal(dAtA []byte) error {
 	l := len(dAtA)
@@ -1064,7 +1348,9 @@ func (m *EventUpdateCommission) Unmarshal(dAtA []byte) error {
 			if postIndex > l {
 				return io.ErrUnexpectedEOF
 			}
-			m.Commission = string(dAtA[iNdEx:postIndex])
+			if err := m.Commission.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
 			iNdEx = postIndex
 		default:
 			iNdEx = preIndex