+++ x/stakers/types/message_create_staker.go
@@ -2,39 +2,37 @@ package types
 
 import (
 	"cosmossdk.io/errors"
+	"github.com/KYVENetwork/chain/util"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	errorsTypes "github.com/cosmos/cosmos-sdk/types/errors"
 )
 
-const TypeMsgCreateStaker = "create_staker"
-
 var _ sdk.Msg = &MsgCreateStaker{}
-
-func (msg *MsgCreateStaker) Route() string {
-	return RouterKey
-}
-
-func (msg *MsgCreateStaker) Type() string {
-	return TypeMsgCreateStaker
-}
 
 func (msg *MsgCreateStaker) GetSigners() []sdk.AccAddress {
 	creator, err := sdk.AccAddressFromBech32(msg.Creator)
 	if err != nil {
 		panic(err)
 	}
+
 	return []sdk.AccAddress{creator}
 }
 
-func (msg *MsgCreateStaker) GetSignBytes() []byte {
-	bz := ModuleCdc.MustMarshalJSON(msg)
-	return sdk.MustSortJSON(bz)
-}
-
 func (msg *MsgCreateStaker) ValidateBasic() error {
-	_, err := sdk.AccAddressFromBech32(msg.Creator)
-	if err != nil {
-		return errors.Wrapf(errorsTypes.ErrInvalidAddress, "invalid creator address (%s)", err)
+	if _, err := sdk.AccAddressFromBech32(msg.Creator); err != nil {
+		return errors.Wrapf(errorsTypes.ErrInvalidAddress, "invalid creator address: %s", err)
 	}
+
+	if util.ValidateNumber(msg.Amount) != nil {
+		return errors.Wrapf(errorsTypes.ErrInvalidRequest, "invalid amount")
+	}
+
+	if msg.Commission.IsNil() {
+		msg.Commission = DefaultCommission
+	}
+	if util.ValidatePercentage(msg.Commission) != nil {
+		return errors.Wrapf(errorsTypes.ErrInvalidRequest, "invalid commission")
+	}
+
 	return nil
 }